requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai
sentence-transformers
faiss-cpu
pypdf2
requests
python-dotenv
asyncio